requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
colorama==0.3.9
cryptography==2.3
Cython==0.28.5
DateTime==4.2
Flask==1.0.2
Flask-Caching==1.4.0
Flask-SQLAlchemy==2.3.2
fuzzywuzzy==0.16.0
gevent==1.3.6
greenlet==0.4.14
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.13
numpy==1.15.0
pandas==0.23.4
pycparser==2.18
python-dateutil==2.7.3
python-decouple==3.1
python-dotenv==0.9.1
python-levenshtein==0.12.0
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.11
Werkzeug==0.14.1
zope.interface==4.5.0
raven[flask]==6.9.0
sentry-sdk[flask]==0.3.7